+++ src/main/java/io/hhplus/tdd/repository/PointHistoryRepository.java
@@ -1,0 +1,8 @@
+package io.hhplus.tdd.repository;
+
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface PointHistoryRepository {
+
+}
