+++ src/main/java/io/hhplus/tdd/repository/PointHistoryRepositoryImpl.java
@@ -1,0 +1,5 @@
+package io.hhplus.tdd.repository;
+
+public class PointHistoryRepositoryImpl implements PointHistoryRepository {
+
+}
