+++ src/main/java/io/hhplus/tdd/repository/UserPointRepositoryImpl.java
@@ -1,0 +1,7 @@
+package io.hhplus.tdd.repository;
+
+import lombok.RequiredArgsConstructor;
+
+public class UserPointRepositoryImpl implements UserPointRepository {
+
+}
